WorldBuilding
I really need to do smaller commits, but here is the summary: I made another tile map for the new assets we had, I found a building tilemap and I implemented and organized that, I worked on the map (the maze can be changed), and I found other inside pixel art for the inside scenes. I increased the size of the character and I lowered the brightness of the houses.